+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
-## [Unreleased](https://github.com/schemathesis/schemathesis/compare/v4.24.1...HEAD) - TBD
+## [Unreleased](https://github.com/schemathesis/schemathesis/compare/v4.24.2...HEAD) - TBD
+
+## [4.24.2](https://github.com/schemathesis/schemathesis/compare/v4.24.1...v4.24.2) - 2026-07-22
 
 ### :bug: Fixed